requirements.txt
asgiref==3.7.1
certifi==2024.12.14
charset-normalizer==3.3.2
Django==5.0.1
google-cloud-storage==2.10.0
gunicorn==21.2.0
idna==3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.1.0
whitenoise==6.6.0
